LayoutTests/inspector/audit/audit-test-case.html

<!doctype html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="./resources/audit-test-fixtures.js"></script>
<script>
function test()
{
    let suite = InspectorTest.createAsyncSuite("AuditTestCase");

    suite.addTestCase({
        name: "Test functions must be asynchronous.",
        description: "AuditTestCase should throw an exception when instantiated with a non-async function.",
        async test() {
            InspectorTest.expectException(() => {
                new WI.AuditTestCase(new testSuiteFixture1, "fakeTest2", () => []);
            });
        }
    })

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Test for the AudtTestCase.</p>
</body>
</html>
